requirements.txt
discord.py==1.7.3
hashlib
asyncio
bs4==0.0.1
html5lib==1.0.1
requests==2.21.0
urllib3==1.24.1
pynacl==1.3.0
pycparser
ffmpeg
cffi==1.12.2
youtube-dl
opuslib
logger
Pillow
discord-py-slash-command
pafy
